feat: Cập nhật chức năng quản lý sản phẩm và hình ảnh
- Thêm API xóa nhiều sản phẩm cùng lúc (soft delete) trong ProductController.
- Cập nhật API tạo sản phẩm mới để hỗ trợ việc tạo sản phẩm với nhiều biến thể.
- Cải thiện các phương thức trong ProductService để xử lý việc xóa nhiều sản phẩm và tạo sản phẩm với biến thể.
- Cập nhật các DTO liên quan để hỗ trợ các thay đổi mới.
- Cải thiện các phương thức trong ProductImageService và ProductImageRepository để sử dụng kiểu dữ liệu Long cho ID biến thể.
- Cập nhật các entity liên quan để đảm bảo tính nhất quán trong việc sử dụng ID kiểu Long.

+++ src/main/java/iuh/fit/supermarket/repository/ProductImageRepository.java
@@ -33,7 +33,7 @@ public interface ProductImageRepository extends JpaRepository<ProductImage, Inte
      * @return danh sách hình ảnh của biến thể
      */
     @Query("SELECT pi FROM ProductImage pi WHERE pi.variant.variantId = :variantId ORDER BY pi.sortOrder ASC, pi.createdAt ASC")
-    List<ProductImage> findByVariantIdOrderBySortOrder(@Param("variantId") Integer variantId);
+    List<ProductImage> findByVariantIdOrderBySortOrder(@Param("variantId") Long variantId);
 
     /**
      * Tìm hình ảnh chính của sản phẩm (sortOrder = 0 hoặc thấp nhất)
@@ -68,7 +68,7 @@ public interface ProductImageRepository extends JpaRepository<ProductImage, Inte
      * @return số lượng hình ảnh
      */
     @Query("SELECT COUNT(pi) FROM ProductImage pi WHERE pi.variant.variantId = :variantId")
-    long countByVariantId(@Param("variantId") Integer variantId);
+    long countByVariantId(@Param("variantId") Long variantId);
 
     /**
      * Xóa tất cả hình ảnh của một sản phẩm
@@ -88,7 +88,7 @@ public interface ProductImageRepository extends JpaRepository<ProductImage, Inte
     @Modifying
     @Transactional
     @Query("DELETE FROM ProductImage pi WHERE pi.variant.variantId = :variantId")
-    void deleteByVariantId(@Param("variantId") Integer variantId);
+    void deleteByVariantId(@Param("variantId") Long variantId);
 
     /**
      * Tìm hình ảnh chung của sản phẩm (không thuộc về biến thể nào)
@@ -115,5 +115,5 @@ public interface ProductImageRepository extends JpaRepository<ProductImage, Inte
      * @return thứ tự sắp xếp lớn nhất
      */
     @Query("SELECT COALESCE(MAX(pi.sortOrder), 0) FROM ProductImage pi WHERE pi.variant.variantId = :variantId")
-    Integer findMaxSortOrderByVariantId(@Param("variantId") Integer variantId);
+    Integer findMaxSortOrderByVariantId(@Param("variantId") Long variantId);
 }
